backend/requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
django-jazzmin
Pillow
python-dotenv
gunicorn
whitenoise
dj-database-url
psycopg2-binary
